Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

The changes look sensible, but you've mixed in the renaming of prices to products. Would be good to move that rename to a separate patch. Would be nicer and easier to see exactly what was changed just for the cache that way.

Also, I really want to get rid of that hardcoded regions dict. We have a place to query it from, we should use it, but that be more related to the product rename than the cache.

Patch-set: 4
Label: Code-Review=-1