Update patch set 3
Patch Set 3: Code-Review-1

I'm not sure what the exact use case for this is.

From what I remember it was to let us spin up an additional distil collector service which would only collect up to a given point.

The issue I have is that for one, we don't have any good deployment mechanism to do that and then clean up the second process and second conf. Also when the collection does catch up, this process will sit there running but do nothing until stopped.

We may want to consider a better way to bring up additional temporary collectors that stop themselves, or maybe devise some better mechanism so we can easily add more collectors, and safely remove them without leaving behind locked projects. That way we treat our collectors as a pool of collectors that we add new ones too as we need more performance.

In general, this feature does what is says, but effectively it leaves behind a zombie process when it does reach the end.

Patch-set: 3
Label: Code-Review=-1