Update patch set 5
Patch Set 5: Code-Review-1

The code looks good, and makes sense, my only concern is that the resource types are not returned as part of the details, which means any consumer of this API has to have a mapping of products to service type. This could be a pain.

Plus, the detail=false case returns only a total cost. It might be better to actually return the type grouping of costs, and a total. So that when detail=false you get a cost per type without detailed per resource info.

I would prefer not to hold this patch back, but if we do decide to add some resource type mappings into these APIs then we should do that now rather than later and break/change features once merged (although backwards compatibility isn't yet really a problem :P).

Patch-set: 5
Label: Code-Review=-1